+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-28T13:23:33.901Z",
+    "2026-04-28T16:26:12.171Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {