+++ packages/orq-rc/RELEASES.md
@@ -8699,3 +8699,13 @@ Based on:
 - [typescript v4.9.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.7 - packages/orq-rc
+
+## 2026-04-28 16:29:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.8 - packages/orq-rc